Fixed PR feedback and ability to save to file.
HOSTING-1133

Removed the + from regexp to validate env vars.

Added a admin permission check for validating the credential.

Any UI dropdown no longer uses the credential id UUID as a value. A public id
is created and is used instead. This is auto created when the KSM credential
is created.

Both freestyle and pipeline redact secrets if displayed in the console.

Removed the HelloWorld annoation.

Switched the builder from build step to a build wrapper. Alot nicer.

Add the ability to save secrets to a file. The file are removed when the build finishes.