Fix problem with test due to Linux/Window difference
Linux will add single quotes around a echo'd value ( '****') and
Windows will not ( **** ). This causes a problem with the test.

Hardcodeed the project name instead of using Messages resource. There
was a problem there with class not being found. Revisit if we do
i18n.

Also setup environment to match Jenkin's build env. Need to to tweak
the pom.xml to get project to build.