Update SDK, fix cert validation, and default radio button
Update the SDK to 16.2.1. This fixes a problem where notation would
not find the field if a label was used.

The certificate validation button would always return an error. This
was due to the hostname entry box being below the validation button.
When passing values it would not include the hostname. Move the
hostname entry box above the validation button.

When adding a new secret, no default radio button is checked. Changed
the code so the a new secret would default to the Env checbox.

Updated the README.md