Fix skipping SSL
There were two problems. The first was initializing a token would not
skip SSL certification. It was not coded to use the skip SSL flag.
The second was testing a credential always resulted in using SSL
certification even if the box was checked. Was using the wrong, non-existing,
field in the validation call.

Also add the SSL flag to log messages. And removed some debug messages
that were not needed.