Update KSM SDK to 16.3.6 for entropy test
KSM-298

On some system, mostly Linux, if the system does not have a lot of
activity there is a chance of low entropy. Entropy is used by
SecrueRandom and it will block until enough exists.

KSM SDK 16.3.6 has a check and will throw an exception if not enough
entropy exists.

    Slow SecureRandom detected!  Install one of the following entropy
    sources to improve speed of random number generator
    on your platform: 'haveged' or 'rng-tools'

Also removed the plugins handling of host aliases (ie US, EU, AU,..)
and now uses the KSM SDK's handling.